+++ src/constants/index.ts
@@ -1,12 +1,15 @@
+export interface Option {
+  value: string
+  label: string
+}
+
 export const paymentTypeOptions = [
   { label: '현금', value: 'cash' },
   { label: '카드', value: 'card' },
 ]
 
-export const dailyResultOptions = ['income', 'expense', 'revenue']
-
-export const dailyResultCategory: { [key: string]: string } = {
-  income: '수입',
-  expense: '지출',
-  revenue: '합계',
-}
+export const dailyResultOptions = [
+  { value: 'income', label: '수입' },
+  { value: 'expense', label: '지출' },
+  { value: 'revenue', label: '합계' },
+]
